+++ mimyria/scripts/apt-from-efield-derivative.py
@@ -4,7 +4,8 @@ import argparse
 from ase.io import iread, write
 
 from mimyria.postprocess import calculate_apt_from_field
-from mimyria.io import apt_flatten
+from mimyria.io import open_wrapper, apt_flatten
+import mimyria.args as common_args
 
 
 def main(argv=None):
@@ -17,12 +18,17 @@ def main(argv=None):
     parser.add_argument('--field_strength', type=float, default=5e-4, help='Strength of the applied electric field (default: 5e-4 atomic units)', required=False)
     parser.add_argument('--out', type=str, default='apt.xyz', help='Output file where to write the APTs (default: apt.xyz)')
 
+    # provides atom_relabel
+    common_args.atom_relabel(parser)
+
     for d in displacements:
         dx = d.replace('p', '+').replace('m', '-')
         parser.add_argument(f'--f{d}', type=str, default=f'field_{d}-frc-1.xyz',
                             help=f'Forces for efield in {dx} direction (default: field_{d}-frc-1.xyz')
 
     args = parser.parse_args(argv)
+
+    atom_relabel = common_args.parse_atom_relabel(args.atom_relabel)
 
     ###############################################
 
@@ -35,7 +41,17 @@ def main(argv=None):
         for d in displacements:
             files[d] = getattr(args, f'f{d}')
 
-    iterators = {key: iread(path) for key, path in files.items()}
+    streams = {
+        key: open_wrapper(path, atom_relabel=atom_relabel)
+        for key, path in files.items()
+    }
+
+    iterators = {
+        key: iread(fin, format=getattr(fin, 'ase_format'))
+        for key, fin in streams.items()
+    }
+
+    fout = open_wrapper(args.out, 'w')
 
     frames2write = []
     for step, frames in enumerate(zip(*iterators.values())):
@@ -51,7 +67,7 @@ def main(argv=None):
 
         print(f'Processed frame {step}')
 
-    write(args.out, frames2write)
+    write(fout, frames2write)
 
 
 if __name__ == "__main__":