+++ mimyria/scripts/pgt-from-efield-derivative.py
@@ -1,10 +1,12 @@
 #!/bin/env python3
 
 import argparse
+import sys
 from ase.io import iread, write
 
 from mimyria.postprocess import calculate_pgt_from_field
 from mimyria.io import open_wrapper, pgt_flatten
+import mimyria.args as common_args
 
 
 def main(argv=None):
@@ -28,6 +30,9 @@ def main(argv=None):
         help='Prints additional memory and timing information at each step'
     )
 
+    # provides atom_relabel
+    common_args.atom_relabel(parser)
+
     parser.add_argument('--out', type=str, required=True)
 
     for d in displacements:
@@ -38,6 +43,8 @@ def main(argv=None):
     args = parser.parse_args(argv)
 
     normalized_field = args.field_mode == 'normalized'
+
+    atom_relabel = common_args.parse_atom_relabel(args.atom_relabel)
 
     ###############################################
 
@@ -51,7 +58,7 @@ def main(argv=None):
             files[d] = getattr(args, f'f{d}')
 
     streams = {
-        key: open_wrapper(path)
+        key: open_wrapper(path, atom_relabel=atom_relabel)
         for key, path in files.items()
     }
 
@@ -60,29 +67,39 @@ def main(argv=None):
         for key, fin in streams.items()
     }
 
-    frames2write = []
-    for step, frames in enumerate(zip(*iterators.values())):
-        atoms = dict(zip(iterators.keys(), frames))
+    fout = open_wrapper(args.out, 'w')
 
-        # NOTE
-        # this gives the PGT in e^2 a_0 / (Eh)
-        # convert to e^2 a_0^2 / (Eh * Angstrom) instead by convention,
-        # since when multiplying by the velocity
-        # (Angstrom / fs), alpha is directly obtained in atomic units.
-        pgts = calculate_pgt_from_field(atoms,
-                                        args.field_strength,
-                                        normalized_field,
-                                        args.enforce_force_sum_rule,
-                                        ) * 1.88973
-        print(pgts.shape)
+    try:
+        frames2write = []
+        for step, frames in enumerate(zip(*iterators.values())):
+            atoms = dict(zip(iterators.keys(), frames))
 
-        trajectory = atoms['pos']
-        trajectory.arrays['pgt'] = pgt_flatten(pgts)
-        frames2write.append(trajectory)
+            # NOTE
+            # this gives the PGT in e^2 a_0 / (Eh)
+            # convert to e^2 a_0^2 / (Eh * Angstrom) instead by convention,
+            # since when multiplying by the velocity
+            # (Angstrom / fs), alpha is directly obtained in atomic units.
+            pgts = calculate_pgt_from_field(atoms,
+                                            args.field_strength,
+                                            normalized_field,
+                                            args.enforce_force_sum_rule,
+                                            ) * 1.88973
+            print(pgts.shape)
 
-        print(f'Processed frame {step}')
+            trajectory = atoms['pos']
+            trajectory.arrays['pgt'] = pgt_flatten(pgts)
+            frames2write.append(trajectory)
 
-    write(args.out, frames2write)
+            print(f'Processed frame {step}')
+
+        write(fout, frames2write)
+
+    except KeyError as e:
+        print(f'ERROR: Processing failed, caught KeyError: {e}', file=sys.stderr)
+        print('NOTE: Usually this is due to atom labels that are not '
+              'directly translatable to element names')
+
+        exit(1)
 
 
 if __name__ == "__main__":